[Broker, Client] Implemented subscription phase
The subscription phase (SUB and SUBACK), only related to the exchanged messages, has been implemented in both broker and client nodes; also, in the definitions.h file an array with the topic names has been added.
However, notice that this step of implementation does not include a proper management of the subscriptions on the broker side; this will be provided in the next commits.

Signed-off-by: Filippo Testa <filippo2.testa@mail.polimi.it>