[Broker] Added subscription management
On the Broker node, proper data structure have been added in order to register and maintain the subscriptions with the related QoS received from the client nodes.
The function responsible for registering the preferences, read the payload from the received message and update the data structures accordingly.
Moreover, a simple error handling mechanism has been implemented and a support function is provided, only for displaying purposes.

Signed-off-by: Filippo Testa <filippo2.testa@mail.polimi.it>