[Client] Implemented Publish management
The client node features new data structure used to store information about publish topic and QoS, subscribed topics and relates QoS levels and, finally, to ensure that the subscription was successful.
A new task has been added with the purpose retransmit a PUB if case it is not ACK from the Broker while two different functions, respectively, decide the publish topic (and QoS) and the topics to subscribe to.
Finally, small modifications have been applied to existing tasks to achieve a seamless integration of the newly implemented PUB phase and to add some missing checks.

Signed-off-by: Filippo Testa <filippo2.testa@mail.polimi.it>